[Setup] Pocket Base 인텔칩셋 버전으로 Re-Setup
