--- backend/pb_migrations/1707818253_created_companies.js
@@ -1,41 +1,0 @@
-/// <reference path="../pb_data/types.d.ts" />
-migrate((db) => {
-  const collection = new Collection({
-    "id": "ux43xcy4ubwu9pw",
-    "created": "2024-02-13 09:57:33.411Z",
-    "updated": "2024-02-13 09:57:33.411Z",
-    "name": "companies",
-    "type": "base",
-    "system": false,
-    "schema": [
-      {
-        "system": false,
-        "id": "jfcecglb",
-        "name": "name",
-        "type": "text",
-        "required": false,
-        "presentable": false,
-        "unique": false,
-        "options": {
-          "min": null,
-          "max": null,
-          "pattern": ""
-        }
-      }
-    ],
-    "indexes": [],
-    "listRule": null,
-    "viewRule": null,
-    "createRule": null,
-    "updateRule": null,
-    "deleteRule": null,
-    "options": {}
-  });
-
-  return Dao(db).saveCollection(collection);
-}, (db) => {
-  const dao = new Dao(db);
-  const collection = dao.findCollectionByNameOrId("ux43xcy4ubwu9pw");
-
-  return dao.deleteCollection(collection);
-})
