--- backend/pb_migrations/1707818323_created_products.js
@@ -1,41 +1,0 @@
-/// <reference path="../pb_data/types.d.ts" />
-migrate((db) => {
-  const collection = new Collection({
-    "id": "c2zigxervwxtnp2",
-    "created": "2024-02-13 09:58:43.928Z",
-    "updated": "2024-02-13 09:58:43.928Z",
-    "name": "products",
-    "type": "base",
-    "system": false,
-    "schema": [
-      {
-        "system": false,
-        "id": "lsaubl0y",
-        "name": "name",
-        "type": "text",
-        "required": false,
-        "presentable": false,
-        "unique": false,
-        "options": {
-          "min": null,
-          "max": null,
-          "pattern": ""
-        }
-      }
-    ],
-    "indexes": [],
-    "listRule": null,
-    "viewRule": null,
-    "createRule": null,
-    "updateRule": null,
-    "deleteRule": null,
-    "options": {}
-  });
-
-  return Dao(db).saveCollection(collection);
-}, (db) => {
-  const dao = new Dao(db);
-  const collection = dao.findCollectionByNameOrId("c2zigxervwxtnp2");
-
-  return dao.deleteCollection(collection);
-})
