--- backend/pb_migrations/1707869178_updated_users.js
@@ -1,58 +1,0 @@
-/// <reference path="../pb_data/types.d.ts" />
-migrate((db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("_pb_users_auth_")
-
-  // remove
-  collection.schema.removeField("users_name")
-
-  // remove
-  collection.schema.removeField("users_avatar")
-
-  return dao.saveCollection(collection)
-}, (db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("_pb_users_auth_")
-
-  // add
-  collection.schema.addField(new SchemaField({
-    "system": false,
-    "id": "users_name",
-    "name": "name",
-    "type": "text",
-    "required": false,
-    "presentable": false,
-    "unique": false,
-    "options": {
-      "min": null,
-      "max": null,
-      "pattern": ""
-    }
-  }))
-
-  // add
-  collection.schema.addField(new SchemaField({
-    "system": false,
-    "id": "users_avatar",
-    "name": "avatar",
-    "type": "file",
-    "required": false,
-    "presentable": false,
-    "unique": false,
-    "options": {
-      "mimeTypes": [
-        "image/jpeg",
-        "image/png",
-        "image/svg+xml",
-        "image/gif",
-        "image/webp"
-      ],
-      "thumbs": null,
-      "maxSelect": 1,
-      "maxSize": 5242880,
-      "protected": false
-    }
-  }))
-
-  return dao.saveCollection(collection)
-})
