--- backend/pb_migrations/1707869217_updated_users.js
@@ -1,24 +1,0 @@
-/// <reference path="../pb_data/types.d.ts" />
-migrate((db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("_pb_users_auth_")
-
-  collection.listRule = null
-  collection.viewRule = null
-  collection.createRule = null
-  collection.updateRule = null
-  collection.deleteRule = null
-
-  return dao.saveCollection(collection)
-}, (db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("_pb_users_auth_")
-
-  collection.listRule = "id = @request.auth.id"
-  collection.viewRule = "id = @request.auth.id"
-  collection.createRule = ""
-  collection.updateRule = "id = @request.auth.id"
-  collection.deleteRule = "id = @request.auth.id"
-
-  return dao.saveCollection(collection)
-})
