--- backend/pb_migrations/1707869238_updated_companies.js
@@ -1,24 +1,0 @@
-/// <reference path="../pb_data/types.d.ts" />
-migrate((db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("ux43xcy4ubwu9pw")
-
-  collection.listRule = ""
-  collection.viewRule = ""
-  collection.createRule = ""
-  collection.updateRule = ""
-  collection.deleteRule = ""
-
-  return dao.saveCollection(collection)
-}, (db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("ux43xcy4ubwu9pw")
-
-  collection.listRule = null
-  collection.viewRule = null
-  collection.createRule = null
-  collection.updateRule = null
-  collection.deleteRule = null
-
-  return dao.saveCollection(collection)
-})
