--- backend/pb_migrations/1707869276_updated_users.js
@@ -1,33 +1,0 @@
-/// <reference path="../pb_data/types.d.ts" />
-migrate((db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("_pb_users_auth_")
-
-  // add
-  collection.schema.addField(new SchemaField({
-    "system": false,
-    "id": "qtg6wd4p",
-    "name": "companies",
-    "type": "relation",
-    "required": false,
-    "presentable": false,
-    "unique": false,
-    "options": {
-      "collectionId": "ux43xcy4ubwu9pw",
-      "cascadeDelete": false,
-      "minSelect": null,
-      "maxSelect": null,
-      "displayFields": null
-    }
-  }))
-
-  return dao.saveCollection(collection)
-}, (db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("_pb_users_auth_")
-
-  // remove
-  collection.schema.removeField("qtg6wd4p")
-
-  return dao.saveCollection(collection)
-})
