--- backend/pb_migrations/1707882787_updated_users.js
@@ -1,44 +1,0 @@
-/// <reference path="../pb_data/types.d.ts" />
-migrate((db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("_pb_users_auth_")
-
-  // update
-  collection.schema.addField(new SchemaField({
-    "system": false,
-    "id": "rngxsrul",
-    "name": "phone_number",
-    "type": "text",
-    "required": false,
-    "presentable": false,
-    "unique": false,
-    "options": {
-      "min": 13,
-      "max": 13,
-      "pattern": "^[0-9]{3}-([0-9]{3}|[0-9]{4})-[0-9]{4}$"
-    }
-  }))
-
-  return dao.saveCollection(collection)
-}, (db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("_pb_users_auth_")
-
-  // update
-  collection.schema.addField(new SchemaField({
-    "system": false,
-    "id": "rngxsrul",
-    "name": "field",
-    "type": "text",
-    "required": false,
-    "presentable": false,
-    "unique": false,
-    "options": {
-      "min": 13,
-      "max": 13,
-      "pattern": "^[0-9]{3}-([0-9]{3}|[0-9]{4})-[0-9]{4}$"
-    }
-  }))
-
-  return dao.saveCollection(collection)
-})
